+++ src/controllers/interaction-controller.ts
@@ -7,7 +7,7 @@ import {
   Log,
   Utils,
 } from "../core";
-import { ChannelState } from "../saveables";
+import { ChannelState } from "../entities";
 import { PlayerJson } from "../types";
 
 export class InteractionController {
