+++ src/entities/channel-state.ts
