+++ src/entities/game/index.ts
