+++ src/entities/game/player-card.ts
