+++ src/entities/game/player.ts
