+++ src/entities/user-state.ts
